fix: add coverage
